decouple acs api util classes to prep for expansion of policy support
