+++ reconcile/acs_rbac.py
@@ -16,10 +16,10 @@ from reconcile.typed_queries.app_interface_vault_settings import (
     get_app_interface_vault_settings,
 )
 from reconcile.utils import gql
-from reconcile.utils.acs_api import (
-    AcsApi,
+from reconcile.utils.acs.rbac import (
+    AcsRbacApi,
     Group,
-    RbacResources,
+    RbacResources
 )
 from reconcile.utils.differ import (
     DiffPair,
@@ -255,7 +255,7 @@ class AcsRbacIntegration(QontractReconcileIntegration[NoParams]):
     def add_rbac_for_role(
         self,
         role: AcsRole,
-        acs: AcsApi,
+        acs: AcsRbacApi,
         admin_access_scope_id: str,
         permission_set_id: str,
         auth_id: str,
@@ -298,7 +298,7 @@ class AcsRbacIntegration(QontractReconcileIntegration[NoParams]):
 
         if not dry_run:
             additions = [
-                AcsApi.GroupAdd(
+                AcsRbacApi.GroupAdd(
                     role_name=role.name,
                     key=a.key,
                     value=a.value,
@@ -317,7 +317,7 @@ class AcsRbacIntegration(QontractReconcileIntegration[NoParams]):
         self,
         to_add: dict[str, AcsRole],
         rbac_api_resources: RbacResources,
-        acs: AcsApi,
+        acs: AcsRbacApi,
         auth_id: str,
         dry_run: bool,
     ) -> list[Exception]:
@@ -351,7 +351,7 @@ class AcsRbacIntegration(QontractReconcileIntegration[NoParams]):
     def delete_rbac_for_role(
         self,
         role: AcsRole,
-        acs: AcsApi,
+        acs: AcsRbacApi,
         access_scope_id: str,
         admin_access_scope_id: str,
         groups: list[Group],
@@ -394,7 +394,7 @@ class AcsRbacIntegration(QontractReconcileIntegration[NoParams]):
         self,
         to_delete: dict[str, AcsRole],
         rbac_api_resources: RbacResources,
-        acs: AcsApi,
+        acs: AcsRbacApi,
         auth_id: str,
         dry_run: bool,
     ) -> list[Exception]:
@@ -430,7 +430,7 @@ class AcsRbacIntegration(QontractReconcileIntegration[NoParams]):
     def update_rbac_for_role(
         self,
         role_diff_pair: DiffPair[AcsRole, AcsRole],
-        acs: AcsApi,
+        acs: AcsRbacApi,
         role_group_mappings: dict[str, dict[str, Group]],
         access_scope_id: str,
         permission_set_id: str,
@@ -465,7 +465,7 @@ class AcsRbacIntegration(QontractReconcileIntegration[NoParams]):
                 for d in diff.delete.values()
             ]
             new = [
-                AcsApi.GroupAdd(
+                AcsRbacApi.GroupAdd(
                     role_name=role_diff_pair.desired.name,
                     key=a.key,
                     value=a.value,
@@ -513,7 +513,7 @@ class AcsRbacIntegration(QontractReconcileIntegration[NoParams]):
         self,
         to_update: dict[str, DiffPair[AcsRole, AcsRole]],
         rbac_api_resources: RbacResources,
-        acs: AcsApi,
+        acs: AcsRbacApi,
         auth_id: str,
         dry_run: bool,
     ) -> list[Exception]:
@@ -558,7 +558,7 @@ class AcsRbacIntegration(QontractReconcileIntegration[NoParams]):
         desired: list[AcsRole],
         current: list[AcsRole],
         rbac_api_resources: RbacResources,
-        acs: AcsApi,
+        acs: AcsRbacApi,
         auth_provider_id: str,
         dry_run: bool,
     ) -> None:
@@ -610,7 +610,7 @@ class AcsRbacIntegration(QontractReconcileIntegration[NoParams]):
 
         desired = self.get_desired_state(gqlapi.query)
 
-        with AcsApi(
+        with AcsRbacApi(
             instance={"url": instance.url, "token": token[instance.credentials.field]}
         ) as acs_api:
             rbac_api_resources = acs_api.get_rbac_resources()
