+++ reconcile/utils/acs/rbac.py
@@ -1,12 +1,7 @@
-from typing import (
-    Any,
-    Optional,
-    Self,
-)
+from typing import Any
 
-import requests
 from pydantic import BaseModel
-
+from reconcile.utils.acs.base import AcsBaseApi
 
 class Role(BaseModel):
     name: str
@@ -121,43 +116,7 @@ def check_len_attributes(attrs: list[Any], api_data: Any) -> None:
             )
 
 
-class AcsApi:
-    def __init__(
-        self,
-        instance: Any,
-        timeout: int = 30,
-    ) -> None:
-        self.base_url = instance["url"]
-        self.token = instance["token"]
-        self.timeout = timeout
-        self.session = requests.Session()
-
-    def __enter__(self) -> Self:
-        return self
-
-    def __exit__(self, exc_type: Any, exc_value: Any, traceback: Any) -> None:
-        self.session.close()
-
-    def generic_request(
-        self, path: str, verb: str, json: Optional[Any] = None
-    ) -> requests.Response:
-        url = f"{self.base_url}{path}"
-        headers = {
-            "Authorization": f"Bearer {self.token}",
-            "Content-Type": "application/json",
-        }
-
-        response = self.session.request(
-            verb,
-            url,
-            headers=headers,
-            json=json,
-            timeout=self.timeout,
-        )
-
-        response.raise_for_status()
-        return response
-
+class AcsRbacApi(AcsBaseApi):
     def get_roles(self) -> list[Role]:
         response = self.generic_request("/v1/roles", "GET")
         return [Role(r) for r in response.json()["roles"]]
